pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["snostr"]

[project]
name = "snostr"
authors = [
  { name="Erik Aronesty", email="erik@q32.com" },
]
description = "Social network scraper for nostr"
keywords = ["nostr"]
urls = { Homepage = "https://github.com/earonesty/snostr" }
readme = "README.md"
requires-python = ">=3.9.0"
dependencies = [
    "snscrape",
    "pynostr",
    "pyyaml",
]
license = {file = "LICENSE"}
classifiers=[
    'Intended Audience :: Developers',
    'License :: OSI Approved :: MIT License',
    'Operating System :: OS Independent',
    'Programming Language :: Python :: 3',
    'Programming Language :: Python :: 3.9',
    'Programming Language :: Python :: 3.10',
    'Programming Language :: Python :: 3.11',
    'Topic :: Internet :: WWW/HTTP',
]

dynamic=["version"]

[tool.setuptools_scm]
write_to = "snostr/_version.py"

[project.scripts]
snostr = "snostr.cli:main"
